cleaning up harmony patches
